+++ src/snippets/manager.ts
@@ -20,10 +20,6 @@ export class SnippetManager {
   private preferComplete: boolean
 
   constructor() {
-    events.on(['TextChanged', 'TextChangedI', 'TextChangedP'], bufnr => {
-      let session = this.getSession(bufnr as number)
-      if (session) session.sychronize()
-    }, null, this.disposables)
     events.on('InsertCharPre', () => {
       // avoid update session when pumvisible
       // Update may cause completion unexpcted terminated.
@@ -50,6 +46,10 @@ export class SnippetManager {
       if (e.affectsConfiguration('suggest') || e.affectsConfiguration('coc.preferences')) {
         this.init()
       }
+    }, null, this.disposables)
+    workspace.onDidChangeTextDocument(e => {
+      let session = this.getSession(e.bufnr)
+      if (session) session.sychronize()
     }, null, this.disposables)
   }